requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
Flask-WTF==1.1.1
pyOpenSSL==23.2.0
